# Add visibility for recent API calls