--- a/arangod/GeneralServer/GeneralServer.cpp
+++ b/arangod/GeneralServer/GeneralServer.cpp
@@ -243,4 +243,3 @@ Result GeneralServer::reloadTLS() {
 }
 
 ArangodServer& GeneralServer::server() const { return _feature.server(); }
-