--- a/arangod/RestServer/ApiRecordingFeature.cpp
+++ b/arangod/RestServer/ApiRecordingFeature.cpp
@@ -1,0 +1,42 @@
+////////////////////////////////////////////////////////////////////////////////
+/// DISCLAIMER
+///
+/// Copyright 2014-2025 ArangoDB GmbH, Cologne, Germany
+/// Copyright 2004-2014 triAGENS GmbH, Cologne, Germany
+///
+/// Licensed under the Business Source License 1.1 (the "License");
+/// you may not use this file except in compliance with the License.
+/// You may obtain a copy of the License at
+///
+///     https://github.com/arangodb/arangodb/blob/devel/LICENSE
+///
+/// Unless required by applicable law or agreed to in writing, software
+/// distributed under the License is distributed on an "AS IS" BASIS,
+/// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+/// See the License for the specific language governing permissions and
+/// limitations under the License.
+///
+/// Copyright holder is ArangoDB GmbH, Cologne, Germany
+///
+/// @author Max Neunhoeffer
+////////////////////////////////////////////////////////////////////////////////
+
+#include "ApiRecordingFeature.h"
+
+#include "ApplicationFeatures/ApplicationServer.h"
+#include "ProgramOptions/ProgramOptions.h"
+
+using namespace arangodb::options;
+
+namespace arangodb {
+
+ApiRecordingFeature::ApiRecordingFeature(Server& server)
+    : ArangodFeature{server, *this} {
+  setOptional(false);
+  startsAfter<application_features::GreetingsFeaturePhase>();
+}
+
+void ApiRecordingFeature::collectOptions(
+    std::shared_ptr<ProgramOptions> options) {}
+
+}  // namespace arangodb