--- a/arangod/RestServer/ApiRecordingFeature.h
+++ b/arangod/RestServer/ApiRecordingFeature.h
@@ -106,7 +106,7 @@ class ApiRecordingFeature : public ArangodFeature {
   void cleanupLoop();
 
   // Whether or not to record recent API calls
-  bool _enabled{true};
+  bool _enabled{false};
 
   // Memory limit for one list of ApiCallRecords:
   size_t _memoryPerApiRecordList{100000};