… Documentation/Metrics/arangodb_api_recording_call_time.yaml
… arangod/RestServer/arangod.h
… arangod/RestServer/arangod_includes.h
… arangod/arangoserver.cmake
… lib/ApplicationFeatures/ApplicationServer.cpp
… lib/ApplicationFeatures/ApplicationServer.h
… lib/Containers/AtomicList.h
… lib/Containers/BoundedList.h
… lib/Containers/ResourceManager.h
… lib/Rest/CommonDefines.h
… tests/Basics/ApplicationServerTest.cpp
… tests/Basics/AtomicListTest.cpp