Add visibility for recent API calls.

This is a test implementation for https://github.com/arangodb/documents/pull/140

We add a new API `/_admin/server/api-calls`, which returns an object like this:

```{
  "error": false,
  "code": 200,
  "result": {
    "calls": [
      {
        "timestamp": 1740493198348,
        "requestType": "GET",
        "path": "/_admin/server/api-calls",
        "database": "_system"
      },
      {
        "timestamp": 1740493196246,
        "requestType": "GET",
        "path": "/_admin/metrics/v2",
        "database": "_system"
      },
      ...
    ]
  }
}
```

There are a few command line options to configure the system:

 - `--server.api-call-recording` (default `true`): flag, if api call records are kept.
 - `--server.memory-per-api-call-list` (default 100000 bytes): amount of memory, when a new api call record list is started
 - `--server.number-of-api-call-lists` (default 256): size of ring buffer for api call record lists

This means, that by default, up to approx. 25 MB are reserved to keep these lists around.

There is one histogram metric, measuring the time spent on this API accounting.

I did a throughput test with the `wrk` tool on a single server with the `/_api/version` call. This is a particularly fast call, so that little changes in performance would have a relatively high impact.

I was using this test:

```wrk -c 40 -d 60 -t 40 --latency http://localhost:8529/_api/version```

This measures for 1 minute and does 40 parallel threads, each with its own connection.

Switching on the `recordAPICall` functionality in this PR gives me:

```
wrk -c 40 -d 60 -t 40 --latency http://localhost:8529/_api/version
Running 1m test @ http://localhost:8529/_api/version
  40 threads and 40 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency    65.18us  236.41us   9.97ms   99.43%
    Req/Sec    19.86k   662.87    24.97k    71.51%
  Latency Distribution
     50%   49.00us
     75%   55.00us
     90%   62.00us
     99%   87.00us
  47499618 requests in 1.00m, 23.62GB read
Requests/sec: 790349.76
Transfer/sec:    402.50MB
```

With the recording switched off I get this:

```
wrk -c 40 -d 60 -t 40 --latency http://localhost:8529/_api/version
Running 1m test @ http://localhost:8529/_api/version
  40 threads and 40 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency    48.40us   42.54us   9.62ms   99.73%
    Req/Sec    20.63k   585.05    27.07k    71.64%
  Latency Distribution
     50%   47.00us
     75%   53.00us
     90%   60.00us
     99%   78.00us
  49338902 requests in 1.00m, 24.54GB read
Requests/sec: 820954.27
Transfer/sec:    418.08MB
```

but the measurements fluctuate anyway nearly as much as the difference.

This amounts to a drop in throughput of approx. 3.8 %. So this is measurable, at least for quick requests. For longer requests this is almost certainly not relevant in this way. Here is a measurement with and without the feature for `POST /_api/document/c`, which has to do more work (and locking!), but still no substantial work.

With recording:

```
wrk -c 40 -d 20 -t 40 --latency http://localhost:8529/_api/document/c
 -s wrk.lua
Running 20s test @ http://localhost:8529/_api/document/c
  40 threads and 40 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency   579.30us    2.94ms  47.38ms   98.56%
    Req/Sec     3.77k   376.55     6.80k    91.04%
  Latency Distribution
     50%  240.00us
     75%  342.00us
     90%  464.00us
     99%   12.89ms
  3019319 requests in 20.10s, 1.68GB read
Requests/sec: 150218.94
Transfer/sec:     85.73MB
```

without recording:

```
wrk -c 40 -d 20 -t 40 --latency http://localhost:8529/_api/document/c
 -s wrk.lua
Running 20s test @ http://localhost:8529/_api/document/c
  40 threads and 40 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency   666.65us    3.31ms  60.44ms   98.08%
    Req/Sec     3.73k   397.78     5.53k    89.59%
  Latency Distribution
     50%  244.00us
     75%  347.00us
     90%  475.00us
     99%   17.01ms
  2984264 requests in 20.10s, 1.67GB read
Requests/sec: 148474.22
Transfer/sec:     84.96MB
```

So here the run was even faster with recording, which only shows that the fluctuations in this test are a lot higher than the time spent in API recording.

The metric measures individual calls to `recordAPICall`. I found the following:

```
# HELP arangodb_api_recording_call_time Execution time histogram for API recording calls [ns]
# TYPE arangodb_api_recording_call_time histogram
arangodb_api_recording_call_time_bucket{role="SINGLE",le="62.500000"} 294201
arangodb_api_recording_call_time_bucket{role="SINGLE",le="125.000000"} 6028441
arangodb_api_recording_call_time_bucket{role="SINGLE",le="250.000000"} 29202991
arangodb_api_recording_call_time_bucket{role="SINGLE",le="500.000000"} 46331172
arangodb_api_recording_call_time_bucket{role="SINGLE",le="1000.000000"} 47118035
arangodb_api_recording_call_time_bucket{role="SINGLE",le="2000.000000"} 47397445
arangodb_api_recording_call_time_bucket{role="SINGLE",le="4000.000000"} 47487262
arangodb_api_recording_call_time_bucket{role="SINGLE",le="8000.000000"} 47497626
arangodb_api_recording_call_time_bucket{role="SINGLE",le="+Inf"} 47499654
arangodb_api_recording_call_time_count{role="SINGLE"} 47499654
arangodb_api_recording_call_time_sum{role="SINGLE"} 14166285702.000000
```

(this was for the `GET /_api/version` measurement.

This means that the average runtime over 47M calls was 298 ns (nanoseconds) per call, including the cases in which
it takes longer. However, for 99.2% of the API calls the time spent is below 1us, and the median is below 250 ns.

Here is the output of a more realistiv workload (shell_client unittests on a single server, running for a few minutes):

```
# HELP arangodb_api_recording_call_time Execution time histogram for API recording calls [ns]
# TYPE arangodb_api_recording_call_time histogram
arangodb_api_recording_call_time_bucket{role="SINGLE",le="62.500000"} 73743
arangodb_api_recording_call_time_bucket{role="SINGLE",le="125.000000"} 193540
arangodb_api_recording_call_time_bucket{role="SINGLE",le="250.000000"} 313272
arangodb_api_recording_call_time_bucket{role="SINGLE",le="500.000000"} 331815
arangodb_api_recording_call_time_bucket{role="SINGLE",le="1000.000000"} 335757
arangodb_api_recording_call_time_bucket{role="SINGLE",le="2000.000000"} 338266
arangodb_api_recording_call_time_bucket{role="SINGLE",le="4000.000000"} 339505
arangodb_api_recording_call_time_bucket{role="SINGLE",le="8000.000000"} 339592
arangodb_api_recording_call_time_bucket{role="SINGLE",le="+Inf"} 339619
arangodb_api_recording_call_time_count{role="SINGLE"} 339619
arangodb_api_recording_call_time_sum{role="SINGLE"} 49905838.000000
```

So here we see that we spent around 147 ns on average for recording an API call and that 99.6% of the calls stay below 2 us.

My conclusion is that this is OK. However, to achieve this it was necessary to do the actual freeing in a separate thread.
This makes design and implementation a bit more complex, but should bring us down to no measurable impact on performance (measured from the outside).

### Scope & Purpose

- [*] :pizza: New feature

### Checklist

- [*] Tests
  - [*] C++ **Unit tests**
- [*] :book: CHANGELOG entry made
- [*] Backports: none planned

#### Related Information

- [*] Design document: https://github.com/arangodb/documents/pull/140