@neunhoef 2025-03-05
Do not merge yet, performance is not where I want it to be with the current solution. We have less overshooting of memory usage, but significant slowdown in the 99%ile. Not good. Will look for improvement tomorrow.